Moved deprecation code to localflavor.deprecation.
