+++ localflavor/be/forms.py
@@ -3,7 +3,7 @@
 from django.forms.fields import RegexField, Select
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
 
 from .be_provinces import PROVINCE_CHOICES
 from .be_regions import REGION_CHOICES
