+++ localflavor/cn/forms.py
@@ -9,7 +9,7 @@ from django.forms.fields import CharField, RegexField, Select
 from django.utils.translation import ugettext_lazy as _
 
 from localflavor.compat import EmptyValueCompatMixin
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
 
 from .cn_provinces import CN_PROVINCE_CHOICES
