+++ localflavor/dk/forms.py
@@ -6,7 +6,7 @@ from django.core.exceptions import ValidationError
 from django.forms import fields, widgets
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
 
 from .dk_municipalities import DK_MUNICIPALITIES
 from .dk_postalcodes import DK_POSTALCODES
