--- localflavor/generic/deprecation.py
@@ -1,2 +1,0 @@
-class RemovedInLocalflavor20Warning(PendingDeprecationWarning):
-    pass
