+++ localflavor/nl/forms.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 from django import forms
 from django.utils import six
 
-from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+from localflavor.deprecation import DeprecatedPhoneNumberFormFieldMixin
 
 from .nl_provinces import PROVINCE_CHOICES
 from .validators import NLPhoneNumberFieldValidator, NLSoFiNumberFieldValidator, NLZipCodeFieldValidator
