+++ localflavor/pk/models.py
@@ -1,7 +1,7 @@
 from django.db.models import CharField
 from django.utils.translation import ugettext_lazy as _
 
-from localflavor.generic.models import DeprecatedPhoneNumberField
+from localflavor.deprecation import DeprecatedPhoneNumberField
 
 from . import forms
 from .pk_states import STATE_CHOICES
