+++ tests/test_deprecated/tests.py
@@ -12,10 +12,10 @@ from localflavor.br.forms import BRPhoneNumberField
 from localflavor.ca.forms import CAPhoneNumberField
 from localflavor.ch.forms import CHPhoneNumberField
 from localflavor.cn.forms import CNCellNumberField, CNPhoneNumberField
+from localflavor.deprecation import RemovedInLocalflavor20Warning
 from localflavor.dk.forms import DKPhoneNumberField
 from localflavor.es.forms import ESPhoneNumberField
 from localflavor.fr.forms import FRPhoneNumberField
-from localflavor.generic.deprecation import RemovedInLocalflavor20Warning
 from localflavor.gr.forms import GRMobilePhoneNumberField, GRPhoneNumberField
 from localflavor.hk.forms import HKPhoneNumberField
 from localflavor.hr.forms import HRPhoneNumberField
